turn vec_map into a dir and split out iters
